added true(er) multilang support
now we are utilising I18n for language strings, that are not unique per form (such as "other", "No questionnaires have been submitted"). this is a bit icky, because we don't use a general magic _
function, but both the localisation features of Forms and I18n. maybe there is a better solution, but ATM i can't think of one.

generally, this fixes #247 and fixes #244 \o/