add ability to make last box a text field for hand written input.
The current way is not very flexible, as last_is_textbox needs to be
specified on the question, so it’s not possible to have multiple (small)
textboxes per row. last_is_textbox takes the width of the box in mm
(Millimeter). Via some hacks this is then passed to the TeX Eval class
which passes the information to OMR in a different way: it gives the
last checkbox a width argument.

Use this guide to know where fixing is required to make this feature
sensible:
Form → TeX: sucks
TeX itself: partly sucks
TeX → OMR:  fine
OMR itself: fine