fix strange hang up
Apparently Rails’ method_missing/respond_to? magic is so fancy that it causes
an endless recursion in the old style (i.e. checking abstract_form first if
method_missing is called; then call super/rails). This was either stopped by
a race condition or nesting limit and it would still work but take 13s. The
new way is to simply have rails execute first and fallback to abstract_form.
This seems to work as well and reduces the runtime to << 1s.