handle asset pipeline more like suggested in the Rails docs.
In production mode this leads to reduced requests because many of the small
files are now included into the single application.js file. The othes get
precompiled, too, although compression should be off for now in our setup.
The compile time is simply too long and the benefits aren’t worth it.