fix saving a new form via ajax to become an editing form
This prevents multiple submissions of the same form which
would result in errors instead of just saving it again. Since
the response is delivered via JSON all templates needed to
have .html.erb explicitely stated, even the subpartial ones.
Otherwise it would result in an template missing error because
Rails was looking for the template.json.