speed up section/question header generation by handling it differently.
Previous method would run complicated CSS selector through jQuery to
determine the correct fields to watch for changes. This turns out to
be very slow (~15s for large form on my pc). The new method uses a
simpler CSS selector and refines the result (~8s). Changes are caught
via a global change-event-listener that is required for undo, anyway.

Basically that gets rid of the “want to cancel long-running script?”
message Firefox produced for me on that form.