add auto correct which improves the OMR great
In our current sample it took down the number of questions a human needs
to look at from ~4500 to ~300 for ~1400 sheets.

Usually a sheet isn’t OMR-able at all, but rather the results become
worse the further it progresses down the page. In other words, questions
on top may still work while the ones on the bottom break. Each checkbox
may be detected in two steps. First, a narrow search is run and iff
that fails a broader one is started. Normally a broad search shouldn’t
be required, so this is an indicator the offset is wrong. If most
checkboxes of a question require a broad search the offset is adjusted
so it works better for the next question.