implement split layout for all hitmes.
Previously only used for the inputs/comments in final check, it
is now used for the job description and metadata as well as
the input/preview on typing, proofreading and combining. Also,
all pages now autosize the comment field.

This should allow for a better usage of current screens, as long
as about 1000px in width are available.