improve speed and progress display for single report creation
The speed improvement is achieved by not starting rake through
a shell again for each course. Rake takes very long to load because
all database connections need to be established and all Rails
libraries loaded.

Also display the common progress status. It counts the persons
(i.e. tutors and profs) needing evaluation instead of the course,
so it’s not skewed for large lectures.